+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -1,18 +1,10 @@
 package seedu.address.logic.parser;
 
-import static java.util.Objects.requireNonNull;
+import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.patient.Age;
-import seedu.address.model.patient.Comment;
-import seedu.address.model.patient.Name;
-import seedu.address.model.patient.PeriodOfStay;
-import seedu.address.model.patient.Phone;
-import seedu.address.model.patient.Temperature;
-import seedu.address.model.patient.TemperatureRange;
-
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
@@ -20,6 +12,8 @@ import seedu.address.model.patient.TemperatureRange;
 public class ParserUtil {
 
     public static final String MESSAGE_INVALID_INDEX = "Index is not a non-zero unsigned integer.";
+    public static final String MESSAGE_INVALID_NUMBER = "Please only enter positive numbers.";
+
     /**
      * Parses {@code oneBasedIndex} into an {@code Index} and returns it. Leading and trailing whitespaces will be
      * trimmed.
@@ -34,106 +28,23 @@ public class ParserUtil {
     }
 
     /**
-     * Parses a {@code String name} into a {@code Name}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code name} is invalid.
+     * Parses {@code integer} into an {@code Integer} and returns it. Leading and trailing whitespaces will be
+     * trimmed.
+     * @throws ParseException if the specified integer is invalid (not non-zero unsigned integer).
      */
-    public static Name parseName(String name) throws ParseException {
-        requireNonNull(name);
-        String trimmedName = name.trim();
-        if (!Name.isValidName(trimmedName)) {
-            throw new ParseException(Name.MESSAGE_CONSTRAINTS);
+    public static Integer parsePositiveInteger(String integer) throws ParseException {
+        String trimmedInteger = integer.trim();
+        if (!StringUtil.isNonZeroUnsignedInteger(trimmedInteger)) {
+            throw new ParseException(MESSAGE_INVALID_NUMBER);
         }
-        return new Name(trimmedName);
+        return Integer.parseInt(trimmedInteger);
     }
 
     /**
-     * Parses a {@code String temperature} into a {@code Temperature}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code temperature} is invalid.
+     * Returns true if none of the prefixes contains empty {@code Optional} values in the given
+     * {@code ArgumentMultimap}.
      */
-    public static Temperature parseTemperature(String temp) throws ParseException {
-        requireNonNull(temp);
-        String trimmedTemp = temp.trim();
-        if (!Temperature.isValidTemperature(trimmedTemp)) {
-            throw new ParseException(Temperature.MESSAGE_CONSTRAINTS);
-        }
-        return new Temperature(trimmedTemp);
-    }
-
-    /**
-     * Parses a {@code String temperature} into a {@code Temperature}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code temperature} is invalid.
-     */
-    public static TemperatureRange parseTemperatureRange(String tempRange) throws ParseException {
-        requireNonNull(tempRange);
-        String trimmedTempRange = tempRange.trim();
-        if (!TemperatureRange.isValidTemperatureRange(trimmedTempRange)) {
-            throw new ParseException(TemperatureRange.MESSAGE_CONSTRAINTS_TEMPERATURERANGE);
-        }
-        return new TemperatureRange(trimmedTempRange);
-    }
-
-    /**
-     * Parses a {@code String period of stay} into a {@code PeriodOfStay}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code period of stay} is invalid.
-     */
-    public static PeriodOfStay parsePeriodOfStay(String periodOfStay) throws ParseException {
-        requireNonNull(periodOfStay);
-        String trimmedPeriodOfStay = periodOfStay.trim();
-        if (!PeriodOfStay.isValidPeriodOfStay(trimmedPeriodOfStay)) {
-            throw new ParseException(PeriodOfStay.MESSAGE_CONSTRAINTS);
-        }
-        return new PeriodOfStay(trimmedPeriodOfStay);
-    }
-
-    /**
-     * Parses a {@code String age} into a {@code Age}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code age} is invalid.
-     */
-    public static Age parseAge(String age) throws ParseException {
-        requireNonNull(age);
-        String trimmedAge = age.trim();
-        if (!Age.isValidAge(trimmedAge)) {
-            throw new ParseException(Age.MESSAGE_CONSTRAINTS);
-        }
-        return new Age(trimmedAge);
-    }
-
-    /**
-     * Parses a {@code String phone} into a {@code Phone}.
-     * Leading and trailing whitespaces will be trimmed.
-     *
-     * @throws ParseException if the given {@code phone} is invalid.
-     */
-    public static Phone parsePhone(String phone) throws ParseException {
-        requireNonNull(phone);
-        String trimmedPhone = phone.trim();
-        if (!Phone.isValidPhone(trimmedPhone)) {
-            throw new ParseException(Phone.MESSAGE_CONSTRAINTS);
-        }
-        return new Phone(trimmedPhone);
-    }
-
-    /**
-     * Parses a {@code String comment} into a {@code Comment}
-     * Leading and trailing whitespaces will be trimmed
-     *
-     * @return an "-" comment if no comment is added
-     */
-    public static Comment parseComment(String comment) {
-        if (comment == null) {
-            return new Comment("-");
-        } else {
-            return new Comment(comment.trim());
-        }
+    public static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
+        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
     }
 }
