+++ src/main/java/seedu/address/logic/parser/patient/SearchPatientCommandParser.java
@@ -4,14 +4,11 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 import static seedu.address.logic.parser.patient.PatientCliSyntax.PREFIX_NAME;
 import static seedu.address.logic.parser.patient.PatientCliSyntax.PREFIX_TEMP_RANGE;
 
-import java.util.stream.Stream;
-
 import seedu.address.logic.commands.patient.SearchPatientCommand;
 import seedu.address.logic.parser.ArgumentMultimap;
 import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
-import seedu.address.logic.parser.Prefix;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
@@ -28,8 +25,8 @@ public class SearchPatientCommandParser implements Parser<SearchPatientCommand> 
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_NAME, PREFIX_TEMP_RANGE);
 
-        if ((!arePrefixesPresent(argMultimap, PREFIX_NAME)
-                && !arePrefixesPresent(argMultimap, PREFIX_TEMP_RANGE))
+        if ((!ParserUtil.arePrefixesPresent(argMultimap, PREFIX_NAME)
+                && !ParserUtil.arePrefixesPresent(argMultimap, PREFIX_TEMP_RANGE))
                 || !argMultimap.getPreamble().isEmpty()) {
             throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, SearchPatientCommand.MESSAGE_USAGE));
         }
@@ -41,22 +38,14 @@ public class SearchPatientCommandParser implements Parser<SearchPatientCommand> 
             String[] commandFields = args.trim().split(" ");
             String nameWithoutPrefix = commandFields[0].substring(2);
             searchPatientDescriptor.setStringName(nameWithoutPrefix);
-            searchPatientDescriptor.setName(ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get()));
+            searchPatientDescriptor.setName(PatientParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get()));
         }
 
         if (argMultimap.getValue(PREFIX_TEMP_RANGE).isPresent()) {
-            searchPatientDescriptor.setTemperatureRange(ParserUtil
-                    .parseTemperatureRange(argMultimap.getValue(PREFIX_TEMP_RANGE).get()));
+            searchPatientDescriptor.setTemperatureRange(
+                    PatientParserUtil.parseTemperatureRange(argMultimap.getValue(PREFIX_TEMP_RANGE).get()));
         }
 
         return new SearchPatientCommand(searchPatientDescriptor);
-    }
-
-    /**
-     * Returns true if none of the prefixes contains empty {@code Optional} values in the given
-     * {@code ArgumentMultimap}.
-     */
-    private static boolean arePrefixesPresent(ArgumentMultimap argumentMultimap, Prefix... prefixes) {
-        return Stream.of(prefixes).allMatch(prefix -> argumentMultimap.getValue(prefix).isPresent());
     }
 }
