+++ src/main/java/seedu/address/logic/parser/room/SearchRoomCommandParser.java
@@ -14,8 +14,8 @@ import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.Prefix;
-import seedu.address.logic.parser.RoomParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.parser.patient.PatientParserUtil;
 import seedu.address.model.patient.Name;
 
 /**
@@ -44,12 +44,12 @@ public class SearchRoomCommandParser implements Parser<SearchRoomCommand> {
         SearchRoomCommand.SearchRoomDescriptor descriptor = new SearchRoomCommand.SearchRoomDescriptor();
 
         if (argMultimap.getValue(PREFIX_NAME).isPresent()) {
-            Name patientName = ParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
+            Name patientName = PatientParserUtil.parseName(argMultimap.getValue(PREFIX_NAME).get());
             descriptor.setPatientName(patientName);
             return new SearchRoomCommand(descriptor);
         }
         //definitely have prefix room number if no prefix name
-        Integer roomNumber = RoomParserUtil.parseRoomNumber(argMultimap.getValue(PREFIX_ROOM_NUMBER).get());
+        Integer roomNumber = ParserUtil.parsePositiveInteger(argMultimap.getValue(PREFIX_ROOM_NUMBER).get());
         descriptor.setRoomNumber(roomNumber);
         return new SearchRoomCommand(descriptor);
